requirements.txt
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
dask==0.13.0
decorator==4.0.10
dlib==19.1.0
enum34==1.1.6
ipython==5.1.0
ipython-genutils==0.1.0
matplotlib==1.5.3
networkx==1.11
numpy==1.11.3
olefile==0.43
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
scikit-image==0.12.3
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
toolz==0.8.2
traitlets==4.3.1
wcwidth==0.1.7
